+++ mullvad-types/src/access_method/mod.rs
@@ -1,0 +1,9 @@
+mod id;
+mod protobuf;
+mod settings;
+mod types;
+
+pub use id::Id;
+pub use protobuf::AccessMethodSetting;
+pub use settings::{Error, Settings};
+pub use types::{AccessMethod, BuiltInAccessMethod};